Move to a mixin pattern for arbitrary extension.
