+++ app/mixins/rel-link-to.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-route-alias/mixins/rel-link-to';
